refactor: extract runtime bootstrap services
