+++ src/dicton/application/__init__.py
@@ -1,0 +1,1 @@
+"""Application services for Dicton."""
