+++ src/dicton/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Bootstrap and composition helpers."""
